Fixed simplemath homework
Take your math contract and modify it so:
The math functions take only one argument, number 2.
Number 1 is kept as a state variable with initial value 0
Each math function executes its operation with number 1 and 2 and stores the result in number 1
Create a function to get the state (number 1)
Create a function that resets the state to zero
Test it and play with it.
Perhaps add more complex logic?